Consider Opera as a Chromium-based browser
This removes Opera from the “other browser viewpoints” list. Instead, two “Opera version shipped” fields are added: one for Opera for Desktop, and another for Opera for Android.

When editing or adding a feature to the site, the corresponding Opera version number is calculated based on the Chromium version that is entered. This happens unless the Opera version number field has been interacted with before the Chromium version was changed. This makes it possible to edit/override the Opera version number if/when needed.

The mock data was updated accordingly.

Ref. #95.