Removed a false compatibility risk title attribute
The title attribute of the compatibility risk color has a static text that does not reflect the state of the world. It looks like it only really matches the green color (which means overall acceptance of the feature).
I am not sure what the text should be for each color, but the lesser evil is to not lie.