Fix editing of entries with multiline values
This fixes an issue (#595) where several of the form fields added in #589
were not marked as multiline entries in the database (i.e. the form fields
could only contain a single line of text). This changes that, and also fixes
an issue with the Web Platform Tests checkbox couldn't be left unchecked.